Update dependencies (#1)
* Updating glm to v1.03

* Updated Lua to v5.5

* updated date in readme copyright license text

* update date

+++ MeshProc/CloseLoopWithPin.cpp
@@ -33,7 +33,7 @@ bool CloseLoopWithPin::Invoke()
 	{
 		center += m_mesh->vertices[i];
 	}
-	center /= m_loop->size();
+	center /= static_cast<float>(m_loop->size());
 
 	center += m_offset;
